Remove port 8080 tunnel
Remove the `LIGHTUP_DATAPLANE_K8S_PORT` config and the tunnel it creates back
to the connect VM. With the HA deployments this fixed config causes port
conflicts on the connect VM. The port currently isn't being used and we can
easily run any `kubectl` commands directly from one of the master nodes.
Alternatively we could make this port configurable or only set it up on a
single master node.