lupmgr preflight checks
Allow the preflight checks from replicated to be run outside of the full installation.
These can then be run by a customer to make sure that their environment is good to go.
If additional checks are needed then they should be addeded in the lightup-replicated
repo.

Also, make sure that root / sudo isn't being used.